Add Trait to generalize button actions
