+++ Core/Lib/ExtendedController/ListBusinessDocument.php
@@ -29,53 +29,7 @@ use FacturaScripts\Dinamic\Lib\BusinessDocumentGenerator;
 abstract class ListBusinessDocument extends ListController
 {
 
-    /**
-     * 
-     * @param string $viewName
-     */
-    protected function addButtonApproveDocument($viewName)
-    {
-        $this->addButton($viewName, [
-            'action' => 'approve-document-same-date',
-            'confirm' => 'true',
-            'icon' => 'fas fa-calendar-check',
-            'label' => 'approve-document-same-date'
-        ]);
-
-        $this->addButton($viewName, [
-            'action' => 'approve-document',
-            'confirm' => 'true',
-            'icon' => 'fas fa-check',
-            'label' => 'approve-document'
-        ]);
-    }
-
-    /**
-     * 
-     * @param string $viewName
-     */
-    protected function addButtonGroupDocument($viewName)
-    {
-        $this->addButton($viewName, [
-            'action' => 'group-document',
-            'icon' => 'fas fa-magic',
-            'label' => 'group-or-split'
-        ]);
-    }
-
-    /**
-     * 
-     * @param string $viewName
-     */
-    protected function addButtonLockInvoice($viewName)
-    {
-        $this->addButton($viewName, [
-            'action' => 'lock-invoice',
-            'confirm' => 'true',
-            'icon' => 'fas fa-lock fa-fw',
-            'label' => 'lock-invoice'
-        ]);
-    }
+    use ListBusinessActionTrait;
 
     /**
      *
@@ -117,51 +71,6 @@ abstract class ListBusinessDocument extends ListController
 
         $currencies = $this->codeModel->all('divisas', 'coddivisa', 'descripcion');
         $this->addFilterSelect($viewName, 'coddivisa', 'currency', 'coddivisa', $currencies);
-    }
-
-    /**
-     * 
-     * @return bool
-     */
-    protected function approveDocumentAction()
-    {
-        if (false === $this->permissions->allowUpdate) {
-            $this->toolBox()->i18nLog()->warning('not-allowed-modify');
-            return true;
-        }
-
-        $codes = $this->request->request->get('code');
-        $model = $this->views[$this->active]->model;
-        if (false === \is_array($codes) || empty($model)) {
-            $this->toolBox()->i18nLog()->warning('no-selected-item');
-            return true;
-        }
-
-        $this->dataBase->beginTransaction();
-        foreach ($codes as $code) {
-            if (false === $model->loadFromCode($code)) {
-                $this->toolBox()->i18nLog()->error('record-not-found');
-                continue;
-            }
-
-            foreach ($model->getAvaliableStatus() as $status) {
-                if (empty($status->generadoc)) {
-                    continue;
-                }
-
-                $model->idestado = $status->idestado;
-                if (false === $model->save()) {
-                    $this->toolBox()->i18nLog()->error('record-save-error');
-                    $this->dataBase->rollback();
-                    return true;
-                }
-            }
-        }
-
-        $this->toolBox()->i18nLog()->notice('record-updated-correctly');
-        $this->dataBase->commit();
-        $model->clear();
-        return true;
     }
 
     /**
@@ -278,111 +187,5 @@ abstract class ListBusinessDocument extends ListController
         }
 
         return parent::execPreviousAction($action);
-    }
-
-    /**
-     * Send the selected codes to the DocumentStitcher controller.
-     *
-     * @return bool
-     */
-    protected function groupDocumentAction()
-    {
-        $codes = $this->request->request->get('code');
-        $model = $this->views[$this->active]->model;
-
-        if (!empty($codes) && $model) {
-            $codes = \implode(',', $codes);
-            $url = 'DocumentStitcher?model=' . $model->modelClassName() . '&codes=' . $codes;
-            $this->redirect($url);
-            return false;
-        }
-
-        $this->toolBox()->i18nLog()->warning('no-selected-item');
-        return true;
-    }
-
-    /**
-     * 
-     * @return bool
-     */
-    protected function lockInvoiceAction()
-    {
-        if (false === $this->permissions->allowUpdate) {
-            $this->toolBox()->i18nLog()->warning('not-allowed-modify');
-            return true;
-        }
-
-        $codes = $this->request->request->get('code');
-        $model = $this->views[$this->active]->model;
-        if (false === \is_array($codes) || empty($model)) {
-            $this->toolBox()->i18nLog()->warning('no-selected-item');
-            return true;
-        }
-
-        $this->dataBase->beginTransaction();
-        foreach ($codes as $code) {
-            if (false === $model->loadFromCode($code)) {
-                $this->toolBox()->i18nLog()->error('record-not-found');
-                continue;
-            }
-
-            foreach ($model->getAvaliableStatus() as $status) {
-                if ($status->editable) {
-                    continue;
-                }
-
-                $model->idestado = $status->idestado;
-                if (false === $model->save()) {
-                    $this->toolBox()->i18nLog()->error('record-save-error');
-                    $this->dataBase->rollback();
-                    return true;
-                }
-            }
-        }
-
-        $this->toolBox()->i18nLog()->notice('record-updated-correctly');
-        $this->dataBase->commit();
-        $model->clear();
-        return true;
-    }
-
-    /**
-     * 
-     * @return bool
-     */
-    protected function paidAction()
-    {
-        if (false === $this->permissions->allowUpdate) {
-            $this->toolBox()->i18nLog()->warning('not-allowed-modify');
-            return true;
-        }
-
-        $codes = $this->request->request->get('code');
-        $model = $this->views[$this->active]->model;
-        if (false === \is_array($codes) || empty($model)) {
-            $this->toolBox()->i18nLog()->warning('no-selected-item');
-            return true;
-        }
-
-        $this->dataBase->beginTransaction();
-        foreach ($codes as $code) {
-            if (false === $model->loadFromCode($code)) {
-                $this->toolBox()->i18nLog()->error('record-not-found');
-                continue;
-            }
-
-            $model->nick = $this->user->nick;
-            $model->pagado = true;
-            if (false === $model->save()) {
-                $this->toolBox()->i18nLog()->error('record-save-error');
-                $this->dataBase->rollback();
-                return true;
-            }
-        }
-
-        $this->toolBox()->i18nLog()->notice('record-updated-correctly');
-        $this->dataBase->commit();
-        $model->clear();
-        return true;
     }
 }
